refactor: config 위치 변경으로 인해 import 변경
