+++ src/main/java/com/tnt/domain/trainee/PtGoal.java
@@ -5,7 +5,7 @@ import static com.tnt.global.error.model.ErrorMessage.PT_GOAL_NULL_TRAINEE_ID;
 import static io.micrometer.common.util.StringUtils.isBlank;
 import static java.util.Objects.requireNonNull;
 
-import com.tnt.global.common.entity.BaseTimeEntity;
+import com.tnt.infrastructure.mysql.BaseTimeEntity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
